requirements.txt
python-dotenv
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-pptx==1.0.0
httpx==0.27.0
pydantic==2.7.3
python-dotenv==1.0.1
